+++ controls/models.py
@@ -7,6 +7,7 @@ from guardian.shortcuts import (assign_perm, get_objects_for_user,
                                 get_perms_for_model, get_user_perms,
                                 get_users_with_perms, remove_perm)
 from simple_history.models import HistoricalRecords
+from jsonfield import JSONField
 
 from .oscal import Catalogs, Catalog
 import uuid
@@ -705,6 +706,9 @@ class Deployment(models.Model):
     updated = models.DateTimeField(auto_now=True, db_index=True)
     uuid = models.UUIDField(default=uuid.uuid4, editable=True, help_text="A UUID (a unique identifier) for the deployment.")
     system = models.ForeignKey('System', related_name='deployments', on_delete=models.CASCADE, blank=True, null=True, help_text="The system associated with the deployment")
+    inventory_items = JSONField(blank=True, null=True,
+        help_text="JSON object representing the inventory items in a deployment.")
+    history = HistoricalRecords(cascade_delete_history=True)
 
     # Notes
     #
@@ -713,6 +717,7 @@ class Deployment(models.Model):
     #    s = System.objects.get(pk=11)
     #    s.deployments.all()
     #    # returns <QuerySet ['ac-2 id=1', 'ac-3 id=2', 'au-2 id=3']>
+    #
 
     def __str__(self):
         return "'%s id=%d'" % (self.name, self.id)
@@ -723,3 +728,43 @@ class Deployment(models.Model):
 
     def get_absolute_url(self):
         return "/systems/%d/deployments" % (self.system.id)
+
+# class AssessmentResults(models.Model):
+#     statement = models.OneToOneField(Statement, related_name="assessment_results",
+#         unique=False, blank=True, null=True, on_delete=models.CASCADE,
+#         help_text="The assessment results details for this statement. Statement must be type 'assessment_results'.")
+#     deployment = models.OneToOneField(Deployment, related_name="assessment_results",
+#         unique=False, blank=True, null=True, on_delete=models.SET_NULL,
+#         help_text="The deployment associated with the assessment results.")
+#     data = JSONField(blank=True, null=True,
+#         help_text="JSON object representing the assessment results items in a deployment.")
+#     ar_type = models.CharField(max_length=150, unique=False, blank=True, null=True,
+#         help_text="Assessment results type.")
+#     generated = models.DateTimeField(db_index=True)
+#     history = HistoricalRecords(cascade_delete_history=True)
+
+#     # Notes
+#     #
+#     # IMPORTANT
+#     #
+#     # JSON data must follow a scheme that is similar to OSCAL.
+#     # Data is assumed to be generated fropm outside of GovReady.
+#     # Data should either be in `data` or `data_binary` field.
+#     # `data_binary` field can hold PDF report or other machine readable format
+#     # Inventory-items must have UUIDs. A UUIDs persits for the life of the instantiaded inventory-item.
+#     #
+#     # The inventory-items in an assessment report can be related to
+#     # UUID of the related inventory-item in the `reference` deployment
+#     # to create a virtual persistence across different instances of
+#     # the "same" assest, such as a virtual database server.
+#     #
+#     # Retrieve System Deployment Inventory
+#     #
+
+#     def __str__(self):
+#         return "<Inventory %s id=%d>" % (self.statement, self.id)
+
+#     def __repr__(self):
+#         # For debugging.
+#         return "<Inventory %s id=%d>" % (self.statement, self.id)
+