+++ siteapp/views.py
@@ -20,7 +20,7 @@ from controls.forms import ImportProjectForm
 from discussion.models import Discussion
 from guidedmodules.models import (Module, ModuleQuestion, ProjectMembership,
                                   Task)
-from controls.models import Element, System
+from controls.models import Element, System, Statement, Poam, Deployment
 from system_settings.models import SystemSettings
 
 from .forms import PortfolioForm, ProjectForm
@@ -555,6 +555,19 @@ def start_app(appver, organization, user, folder, task, q, portfolio):
             object={"object": "element", "id": element.id, "name":element.name},
             user={"id": user.id, "username": user.username}
         )
+        # Add deault deployments to system
+        deployment = Deployment(name="Design", description="Reference system archictecture design", system=system)
+        deployment.save()
+        deployment = Deployment(name="Dev", description="Development environment deployment", system=system)
+        deployment.save()
+        deployment = Deployment(name="Stage", description="Stage/Test environment deployment", system=system)
+        deployment.save()
+        deployment = Deployment(name="Prod", description="Production environment deployment", system=system)
+        deployment.save()
+        # Assign default control catalog
+        # Assign default control profile for org systems
+        # Assign default organization components for a system
+        # Assign default org params
 
         # Add user as the first admin.
         ProjectMembership.objects.create(